+++ src/diagnostics/default_diagnostics.jl
@@ -444,7 +444,10 @@ function add_diagnostics!(
     model::CanopyModel,
     subcomponent::ZhouOptimalLAIModel,
 )
-    append!(diagnostics, ["a0d", "a0a", "pra"])
+    append!(
+        diagnostics,
+        ["a0d", "a0a", "a0c3", "a0c4", "pra", "fc3", "ftr", "fc3g", "fc4g"],
+    )
     return nothing
 end