+++ src/standalone/Vegetation/Canopy.jl
@@ -14,6 +14,7 @@ import LinearAlgebra: I, dot
 using ClimaLand: AbstractRadiativeDrivers, AbstractAtmosphericDrivers
 import ..Parameters as LP
 import Insolation.Parameters as IP
+import Thermodynamics.Parameters as TP
 using Dates
 
 import ClimaLand:
@@ -411,7 +412,6 @@ end
     ZhouOptimalLAIModel{FT}(
         domain,
         toml_dict::CP.ParamDict;
-        optimal_lai_inputs = optimal_lai_static_inputs(domain.space.surface),
         SAI::FT = toml_dict["SAI"],
         RAI::FT = toml_dict["RAI"],
         rooting_depth = clm_rooting_depth(domain.space.surface),
@@ -430,8 +430,6 @@ mirrored into `p.canopy.biomass.area_index.leaf`.
 - `toml_dict`: Parameter dictionary containing optimal LAI parameters
 
 # Keyword Arguments
-- `optimal_lai_inputs`: NamedTuple with the spatially varying inputs of the LAI formulas
-  (GSL, vpd_gs, f0). Default loads from `optimal_lai_static_inputs`.
 - `SAI`: Stem area index (m2/m2), default from toml_dict
 - `RAI`: Root area index (m2/m2), default from toml_dict
 - `rooting_depth`: Rooting depth (m), default from CLM data
@@ -450,7 +448,6 @@ Global Change Biology. https://onlinelibrary.wiley.com/doi/pdf/10.1111/gcb.70125
 function ZhouOptimalLAIModel{FT}(
     domain,
     toml_dict::CP.ParamDict;
-    optimal_lai_inputs = optimal_lai_static_inputs(domain.space.surface),
     SAI::FT = toml_dict["SAI"],
     RAI::FT = toml_dict["RAI"],
     rooting_depth = clm_rooting_depth(domain.space.surface),
@@ -462,14 +459,7 @@ function ZhouOptimalLAIModel{FT}(
     height = toml_dict["canopy_height"],
 ) where {FT <: AbstractFloat}
     parameters = OptimalLAIParameters{FT}(toml_dict)
-    return ZhouOptimalLAIModel{FT}(
-        parameters,
-        optimal_lai_inputs;
-        SAI,
-        RAI,
-        rooting_depth,
-        height,
-    )
+    return ZhouOptimalLAIModel{FT}(parameters; SAI, RAI, rooting_depth, height)
 end
 
 ## Radiative transfer models
@@ -703,6 +693,26 @@ struct CanopyModel{FT, AR, RM, PM, SM, SMSM, PHM, EM, SIFM, BM, B, PSE, D} <:
 end
 
 """
+    check_component_compatibility(photosynthesis, conductance, biomass)
+
+Asserts that the canopy components can be used together: the P-model requires the
+P-model stomatal conductance and vice versa, and `ZhouOptimalLAIModel` requires the
+P-model.
+"""
+function check_component_compatibility(photosynthesis, conductance, biomass)
+    if photosynthesis isa PModel
+        @assert conductance isa PModelConductance "When using PModel for photosynthesis, you must also use PModelConductance for stomatal conductance"
+    end
+    if conductance isa PModelConductance
+        @assert photosynthesis isa PModel "When using PModelConductance for stomatal conductance, you must also use PModel for photosynthesis"
+    end
+    if biomass isa ZhouOptimalLAIModel
+        @assert photosynthesis isa PModel "When using ZhouOptimalLAIModel for biomass, you must also use PModel for photosynthesis"
+    end
+    return nothing
+end
+
+"""
     CanopyModel{FT}(;
         autotrophic_respiration::AbstractAutotrophicRespirationModel{FT},
         radiative_transfer::AbstractRadiationModel{FT},
@@ -745,13 +755,7 @@ function CanopyModel{FT}(;
     },
 ) where {FT, B, PSE}
 
-    if typeof(photosynthesis) <: PModel{FT}
-        @assert typeof(conductance) <: PModelConductance{FT} "When using PModel for photosynthesis, you must also use PModelConductance for stomatal conductance"
-    end
-
-    if typeof(conductance) <: PModelConductance{FT}
-        @assert typeof(photosynthesis) <: PModel{FT} "When using PModelConductance for stomatal conductance, you must also use PModel for photosynthesis"
-    end
+    check_component_compatibility(photosynthesis, conductance, biomass)
 
     args = (
         autotrophic_respiration,
@@ -882,6 +886,7 @@ function CanopyModel{FT}(
         earth_param_set,
         domain,
     )
+    check_component_compatibility(photosynthesis, conductance, biomass)
     return CanopyModel{FT, typeof.(args)...}(args...)
 end
 
@@ -979,6 +984,7 @@ function CanopyModel{FT}(
         earth_param_set,
         domain,
     )
+    check_component_compatibility(photosynthesis, conductance, biomass)
     return CanopyModel{FT, typeof.(args)...}(args...)
 end